% Generated by build script
\name{nih_spending}
\alias{nih_spending}
\title{Get NIH spending by institute/center}
\usage{
nih_spending()
}
\arguments{
\item{fiscal_year}{Fiscal year (default current year)}
\item{limit}{Max results (default 100)}
}
\value{tibble: agency, total_funding, project_count}
\description{Get NIH spending by institute/center}
