% Generated by build script
\name{pubchem_assays}
\alias{pubchem_assays}
\title{Get bioactivity/assay data for a compound}
\usage{
pubchem_assays()
}
\arguments{
\item{cid}{PubChem Compound ID (integer)}
\item{max_results}{Maximum assays to return (default 20)}
}
\value{tibble: aid, assay_name, activity_outcome, target_name}
\description{Get bioactivity/assay data for a compound}
