% Generated by build script
\name{pubchem_compound}
\alias{pubchem_compound}
\title{Look up a compound by name in PubChem}
\usage{
pubchem_compound()
}
\arguments{
\item{name}{Compound name (e.g., "aspirin", "caffeine", "glucose")}
}
\value{tibble: cid, iupac_name, molecular_formula, molecular_weight,}
\description{Look up a compound by name in PubChem}
