% Generated by build script
\name{pubchem_search}
\alias{pubchem_search}
\title{Search PubChem compounds by text query}
\usage{
pubchem_search()
}
\arguments{
\item{query}{Search text (e.g., "aspirin", "anti-inflammatory")}
\item{max_results}{Maximum number of results (default 10)}
}
\value{tibble: cid, iupac_name, molecular_formula, molecular_weight}
\description{Search PubChem compounds by text query}
