% Generated by build script
\name{rxnorm_class}
\alias{rxnorm_class}
\title{Get drug classes for an RxCUI}
\usage{
rxnorm_class()
}
\arguments{
\item{rxcui}{RxNorm concept ID}
\item{source}{Classification source: "ATC", "VA", "MESH", "MEDRT".}
}
\value{tibble: class_id, class_name, class_type, rxcui, drug_name}
\description{Get drug classes for an RxCUI}
