% Generated by build script
\name{rxnorm_related}
\alias{rxnorm_related}
\title{Get all related concepts for a drug}
\usage{
rxnorm_related()
}
\arguments{
\item{rxcui}{RxNorm concept ID}
\item{tty}{Optional term type filter (e.g. "BN" for brand names,}
}
\value{tibble: rxcui, name, tty, synonym}
\description{Get all related concepts for a drug}
